Various doc string cleanups
