+++ cirq/ops/raw_types.py
@@ -367,6 +367,13 @@ class Operation(metaclass=abc.ABCMeta):
 
     @abc.abstractmethod
     def with_qubits(self, *new_qubits: 'cirq.Qid') -> 'cirq.Operation':
+        """Returns the same operation, but applied to different qubits.
+
+        Args:
+            new_qubits: The new qubits to apply the operation to. The order must
+                exactly match the order of qubits returned from the operation's
+                `qubits` property.
+        """
         pass
 
     def transform_qubits(self, func: Callable[[Qid], Qid]) -> 'Operation':
